# docs/openAPI.yaml
openapi: 3.0.3
info:
  title: Decentralized S3-Compatible Storage Gateway API
  version: 1.0.0
  description: |
    S3-compatible API for decentralized storage on IPFS with client-side encryption.
    
    This API provides full S3 compatibility while leveraging:
    - IPFS Cluster for distributed persistence
    - Prolly Trees for efficient indexing
    - Client-side HPKE encryption (keys never leave client devices)
    - BLAKE3/Bao for verified streaming
    - CRDTs for conflict-free collaboration
    
    **Authentication**: OAuth 2.0 with JWT Bearer tokens
    
    **Key Features**:
    - Zero-knowledge storage (gateway sees only encrypted blocks)
    - Automatic deduplication via content-addressing
    - Real-time collaboration with CRDTs
    - Versioning and immutable history
    
  contact:
    name: API Support
    url: https://github.com/functionland/dsn-gateway
    email: hi@fx.land
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: https://gateway.fx.land
    description: Production Gateway
  - url: https://staging.gateway.fx.land
    description: Staging Environment
  - url: http://localhost:9000
    description: Local Development
...
security:
  - oauth2: [storage:read, storage:write]
  - bearerAuth: []
paths:
...
  /{bucket}:
    parameters:
      - $ref: '#/components/parameters/BucketName'
    
    put:
      tags:
        - Bucket Operations
      summary: CreateBucket
      operationId: createBucket
      description: Creates a new bucket with the specified name
      security:
        - oauth2: [storage:write]
      parameters:
        - name: x-amz-acl
          in: header
          description: Canned ACL to apply to the bucket
          schema:
            type: string
            enum: [private, public-read, public-read-write, authenticated-read]
            default: private
      responses:
        '200':
          description: Bucket created successfully
          headers:
            Location:
              schema:
                type: string
              description: URI of the created bucket
        '409':
          $ref: '#/components/responses/BucketAlreadyExists'
        '500':
          $ref: '#/components/responses/InternalError'
    
    head:
      tags:
        - Bucket Operations
      summary: HeadBucket
      operationId: headBucket
      description: Determines if a bucket exists and the user has permission to access it
      security:
        - oauth2: [storage:read]
      responses:
        '200':
          description: Bucket exists and user has access
        '404':
          $ref: '#/components/responses/NoSuchBucket'
        '403':
          $ref: '#/components/responses/AccessDenied'
    
    delete:
      tags:
        - Bucket Operations
      summary: DeleteBucket
      operationId: deleteBucket
      description: Deletes an empty bucket
      security:
        - oauth2: [storage:write]
      responses:
        '204':
          description: Bucket deleted successfully
        '404':
          $ref: '#/components/responses/NoSuchBucket'
        '409':
          description: Bucket not empty
          content:
            application/xml:
              schema:
                $ref: '#/components/schemas/Error'
              example: |
                <?xml version="1.0" encoding="UTF-8"?>
                <Error>
                  <Code>BucketNotEmpty</Code>
                  <Message>The bucket you tried to delete is not empty</Message>
                  <BucketName>my-bucket</BucketName>
                </Error>
    
    get:
      tags:
        - Object Operations
      summary: ListObjectsV2
      operationId: listObjectsV2
      description: |
        Returns objects in a bucket. Uses Prolly Tree indexing for O(log n) performance.
        
        **Performance**: Returns 1,000 keys in <200ms via efficient tree traversal.
      security:
        - oauth2: [storage:read]
      parameters:
        - name: list-type
          in: query
          description: Version of list operation (must be 2 for ListObjectsV2)
          required: true
          schema:
            type: integer
            enum: [2]
        - name: prefix
          in: query
          description: Limits response to keys beginning with the specified prefix
          schema:
            type: string
          example: "photos/2025/"
        - name: delimiter
          in: query
          description: Character used to group keys (typically "/")
          schema:
            type: string
          example: "/"
        - name: max-keys
          in: query
          description: Maximum number of keys to return
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 1000
        - name: continuation-token
          in: query
          description: Token for paginated results (from previous response)
          schema:
            type: string
        - name: start-after
          in: query
          description: Start listing after this specified key
          schema:
            type: string
        - name: fetch-owner
          in: query
          description: Include owner information in response
          schema:
            type: boolean
            default: false
        - name: encoding-type
          in: query
          description: Encoding type for keys in the response
          schema:
            type: string
            enum: [url]
      responses:
        '200':
          description: Successfully retrieved object list
          content:
            application/xml:
              schema:
                $ref: '#/components/schemas/ListBucketResult'
              example: |
                <?xml version="1.0" encoding="UTF-8"?>
                <ListBucketResult xmlns="http://s3.amazonaws.com/doc/2006-03-01/">
                  <Name>my-bucket</Name>
                  <Prefix>photos/</Prefix>
                  <KeyCount>3</KeyCount>
                  <MaxKeys>1000</MaxKeys>
                  <IsTruncated>false</IsTruncated>
                  <Contents>
                    <Key>photos/beach.jpg</Key>
                    <LastModified>2025-12-07T01:42:00.000Z</LastModified>
                    <ETag>"b2419b1e3fd45d596ee22bdf62aaaa2f"</ETag>
                    <Size>2048576</Size>
                    <StorageClass>STANDARD</StorageClass>
                  </Contents>
                  <Contents>
                    <Key>photos/sunset.png</Key>
                    <LastModified>2025-12-07T03:15:00.000Z</LastModified>
                    <ETag>"8f47e10b6a3c9d2e5f1a4b7c3d8e6f9a"</ETag>
                    <Size>1536000</Size>
                    <StorageClass>STANDARD</StorageClass>
                  </Contents>
                  <CommonPrefixes>
                    <Prefix>photos/2024/</Prefix>
                  </CommonPrefixes>
                  <CommonPrefixes>
                    <Prefix>photos/2025/</Prefix>
                  </CommonPrefixes>
                </ListBucketResult>
        '404':
          $ref: '#/components/responses/NoSuchBucket'

  # ============================================================================
  # OBJECT OPERATIONS
  # ============================================================================
...
  /{bucket}?delete:
    parameters:
      - $ref: '#/components/parameters/BucketName'
    
    post:
      tags:
        - Object Operations
      summary: DeleteObjects (Batch Delete)
      operationId: deleteObjects
      description: Deletes multiple objects in a single request (max 1000)
      security:
        - oauth2: [storage:write]
      requestBody:
        required: true
        content:
          application/xml:
            schema:
              type: object
              xml:
                name: Delete
              properties:
                Quiet:
                  type: boolean
                  description: Enable quiet mode (return only errors)
                Object:
                  type: array
                  items:
                    type: object
                    properties:
                      Key:
                        type: string
                      VersionId:
                        type: string
            example: |
              <?xml version="1.0" encoding="UTF-8"?>
              <Delete>
                <Quiet>false</Quiet>
                <Object>
                  <Key>sample1.txt</Key>
                </Object>
                <Object>
                  <Key>sample2.txt</Key>
                </Object>
              </Delete>
      responses:
        '200':
          description: Batch delete completed
          content:
            application/xml:
              schema:
                $ref: '#/components/schemas/DeleteResult'
              example: |
                <?xml version="1.0" encoding="UTF-8"?>
                <DeleteResult xmlns="http://s3.amazonaws.com/doc/2006-03-01/">
                  <Deleted>
                    <Key>sample1.txt</Key>
                  </Deleted>
                  <Error>
                    <Key>sample2.txt</Key>
                    <Code>AccessDenied</Code>
                    <Message>Access Denied</Message>
                  </Error>
                </DeleteResult>

  # ============================================================================
  # MULTIPART UPLOAD
  # ============================================================================
...